move spi test code into separate module
